refactor(checker): route overlap relation guards
