+++ crates/tsz-checker/src/types/utilities/mod.rs
@@ -1,9 +1,11 @@
 //! Parameter type utilities, type construction, and type resolution methods
 //! for `CheckerState`.
 
+pub(crate) mod class_navigation_helpers;
 pub(crate) mod const_enum_eval;
 pub(crate) mod contextual_calls;
 pub(crate) mod core;
 pub(crate) mod cycle_guard;
 pub(crate) mod enum_utils;
+pub(crate) mod overlap_relation_helpers;
 pub(crate) mod return_type;
